Korrigiert Person löschen Funktion

+++ persoenlichkeit_loeschen.php
@@ -54,13 +54,15 @@ if(isset($_GET["nachname"])) {
 
 
 //Alle Verknüpfungen werden gelöscht
-$dbcontroller->deletePersoenlichkeitEpocheOfAPersoenlichkeit($personID);
-$dbcontroller->deletePersoenlichkeitKategorieOfAPersoenlichkeit($personID);
-$dbcontroller->deletePersoenlichkeitLiteraturangabenOfAPersoenlicheit($personID);
-$dbcontroller->deletePersoenlichkeitPersoenlichkeitofAPersoenlichkeit($personID);
-$dbcontroller->deletePersoenlichkeit($personID);
-$dbcontroller->deleteBild($titelbild);
-$dbcontroller->deleteBild($profilbild);
+if($confirmed==1) {
+    $dbcontroller->deletePersoenlichkeitEpocheOfAPersoenlichkeit($personID);
+    $dbcontroller->deletePersoenlichkeitKategorieOfAPersoenlichkeit($personID);
+    $dbcontroller->deletePersoenlichkeitLiteraturangabenOfAPersoenlicheit($personID);
+    $dbcontroller->deletePersoenlichkeitPersoenlichkeitofAPersoenlichkeit($personID);
+    $dbcontroller->deletePersoenlichkeit($personID);
+    $dbcontroller->deleteBild($titelbild);
+    $dbcontroller->deleteBild($profilbild);
+}
